+++ packages-next/admin-ui/src/templates/list.tsx
@@ -1,5 +1,0 @@
-export const listTemplate = (listKey: string) =>
-  `import { getListPage } from '@keystone-next/admin-ui/pages/ListPage';
-
-export default getListPage(${JSON.stringify({ listKey })});
-`;