+++ packages-next/fields/src/types/password/views/index.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 
-import { CellContainer } from '@keystone-next/admin-ui/components';
+import { CellContainer } from '@keystone-next/keystone/admin-ui/components';
 import {
   CardValueComponent,
   CellComponent,